% file MASS/accdeaths.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{accdeaths}
\alias{accdeaths}
\title{
Accidental Deaths in the US 1973-1978
}
\description{
A regular time series giving the monthly totals of accidental
deaths in the USA. The values for first six months of 1979 (p. 326) are 
7798 7406 8363 8460 9217 9316
}
\source{
P. J. Brockwell and R. A. Davis (1991)
\emph{Time Series: Theory and Methods.}
Springer, New York.
}
\keyword{datasets}
