% file MASS/deaths.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{deaths}
\alias{deaths}
\title{
Monthly Deaths from Lung Diseases in the UK
}
\description{
A regular time series giving the monthly deaths from bronchitis,
emphysema and asthma in the UK,
1974-1979, both sexes (\code{deaths}), males (\code{mdeaths}) and females (\code{fdeaths}).


}
\source{
P. J. Diggle (1990)
\emph{Time Series: A Biostatistical Introduction.}
Oxford, table A.3
}
\seealso{
\code{\link{fdeaths}}, \code{\link{mdeaths}}
}
\keyword{datasets}
