\name{diff.ts}
\alias{diff.ts}
\title{diff Method for ts Objects}
\description{
  \code{diff} method for \code{ts} objects.
}
\usage{
diff.ts(x, lag=1, differences=1)
}
\arguments{
  \item{x}{an object of class \code{"ts"} containing the values to be
    differenced.}
  \item{lag}{an integer indicating which lag to use.}
  \item{differences}{an integer indicating the order of the difference.}
}
\seealso{\code{\link{diff}}}
\keyword{ts}
\keyword{arith}
