% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.ts_tokens}
\alias{format.ts_tokens}
\title{Format tree-sitter tokens}
\usage{
\method{format}{ts_tokens}(x, n = 10, ...)
}
\arguments{
\item{x}{ts_tokens object.}

\item{n}{Number of lines, or number of selections to print.}

\item{...}{Ignored.}
}
\value{
Character vector.
}
\description{
Format a ts_tokens object for printing.
}
\details{
This is the engine of \code{\link[=print.ts_tokens]{print.ts_tokens()}}, possibly useful to obtain a
printed representation without doing the actual printing.
}
\examples{
# TODO
}
