% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IGHD_levels}
\alias{IGHD_levels}
\title{Human IGHD genes according to their relative locus location}
\format{
A character vector of length 27
}
\usage{
as.factor(string, levels = TS::IGHD_levels)
}
\description{
A vector that can be used to select and arrange IGHV genes in the correct order.
Data obtained from IMGT
}
\keyword{datasets}
