% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IGHJ_levels}
\alias{IGHJ_levels}
\title{Human IGHJ genes according to their relative locus location}
\format{
A character vector of length 6
}
\usage{
as.factor(string, levels = TS::IGHJ_levels)
}
\description{
A vector that can be used to select and arrange IGHJ genes in the correct order.
Data obtained from IMGT
}
\keyword{datasets}
