% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IGHV_family_levels}
\alias{IGHV_family_levels}
\title{Human IGHV families in ascending numeric order}
\format{
A character vector of length 7
}
\usage{
as.factor(string, levels = TS::IGHV_family_levels)
}
\description{
A vector that can be used to select and arrange IGHV families.
Data obtained from IMGT
}
\keyword{datasets}
