% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IGHV_family_levels_grouped}
\alias{IGHV_family_levels_grouped}
\title{Human IGHV families in ascending numeric order with IGHV6 and 7 grouped}
\format{
A character vector of length 6
}
\usage{
as.factor(string, levels = TS::IGHV_family_levels_grouped)
}
\description{
A vector that can be used to select and arrange IGHV families while IGHV6 and 7 are grouped as they only contain one gene each.
This requires that IGHV_family is set to IGHV6/7 for those families, which is done automatically by TS_plot_IG_usage(compare_what = "IGHV_family_grouped")
Data obtained from IMGT
}
\keyword{datasets}
