% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IGKJ_levels}
\alias{IGKJ_levels}
\title{Human IGKJ genes according to their relative locus location}
\format{
A character vector of length 5
}
\usage{
as.factor(string, levels = TS::IGKJ_levels)
}
\description{
A vector that can be used to select and arrange IGKJ genes in the correct order.
Data obtained from IMGT
}
\keyword{datasets}
