% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IGKV_levels}
\alias{IGKV_levels}
\title{Human IGKV genes according to their relative locus location}
\format{
A character vector of length 76
}
\usage{
as.factor(string, levels = TS::IGKV_levels)
}
\description{
A vector that can be used to select and arrange IGKV genes in the correct order.
Data obtained from IMGT
}
\keyword{datasets}
