% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IGLC_levels}
\alias{IGLC_levels}
\title{Human IGLC genes according to their relative locus location}
\format{
A character vector of length 7
}
\usage{
as.factor(string, levels = TS::IGLC_levels)
}
\description{
A vector that can be used to select and arrange IGLC genes in the correct order.
Data obtained from IMGT
}
\keyword{datasets}
