% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IGLV_family_levels}
\alias{IGLV_family_levels}
\title{Human IGLV families in ascending numeric order}
\format{
A character vector of length 11
}
\usage{
as.factor(string, levels = TS::IGLV_family_levels)
}
\description{
A vector that can be used to select and arrange IGLV families.
Data obtained from IMGT
}
\keyword{datasets}
