% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IGLV_levels}
\alias{IGLV_levels}
\title{Human IGLV genes according to their relative locus location}
\format{
A character vector of length 74
}
\usage{
as.factor(string, levels = TS::IGLV_levels)
}
\description{
A vector that can be used to select and arrange IGLV genes in the correct order.
Data obtained from IMGT
}
\keyword{datasets}
