% file MASS/lh.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{lh}
\alias{lh}
\title{
Luteinizing Hormone in Blood Samples
}
\description{
A regular time series giving the luteinizing hormone in blood
samples at 10mins intervals from a human female, 48 samples. 


}
\source{
P.J. Diggle (1990)
\emph{Time Series: A Biostatistical Introduction.}
Oxford, table A.1, series 3
}
\keyword{datasets}
