% file MASS/nottem.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{nottem}
\alias{nottem}
\title{
Average Monthly Temperatures at Nottingham, 1920-1939
}
\description{
A regular time-series object containing average air temperatures at Nottingham
Castle in degrees F for 20 years.
}
\source{
Anderson, O.D. (1976)
\emph{Time Series Analysis and Forecasting: The Box-Jenkins approach.}
Butterworths. Series R.
}
\keyword{datasets}
