% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ts_tokens}
\alias{print.ts_tokens}
\title{Print tree-sitter tokens}
\usage{
\method{print}{ts_tokens}(x, n = 10, ...)
}
\arguments{
\item{x}{ts_tokens object.}

\item{n}{Number of lines, or number of selections to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Calls \code{\link[=format.ts_tokens]{format.ts_tokens()}} to format the ts_tokens object, writes the
formatted object to the standard output, and returns the original object
invisibly.
}
\examples{
# TODO
}
