% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_usage.R
\name{ram_usage}
\alias{ram_usage}
\title{RAM usage of objects.}
\usage{
ram_usage(number_of_objects = 5)
}
\arguments{
\item{number_of_objects}{Numeric. Number of objects to list. Defaults to 5.}
}
\value{
List of the \code{number_of_objects} objects that uses the most RAM memory.
}
\description{
Lists RAM usage of n objects in the global environment.
}
