\name{sales}
\alias{sales}
\title{
  Sales Data with Leading Indicator.
}
\usage{
data (sales)
}
\description{
  The sales time series \code{sales} and leading indicator
  \code{sales.lead} each contain 150 observations.
  The objects are of class \code{"ts"}.
}
\source{
    The data are given in Box & Jenkins (1976).
  Obtained from the Time Series Data Library at
  \url{http://www-personal.buseco.monash.edu.au/~hyndman/TSDL/}
}

\references{
    G. E. P. Box and G. M. Jenkins (1976): \emph{Time Series
	Analysis, Forecasting and Control}, Holden-Day, San
    Francisco, p. 537.

    P. J. Brockwell and R. A. Davis (1991): \emph{Time Series:
	Theory and Methods}, Second edition, Springer Verlag, NY, pp. 414.
}

\keyword{datasets}


