% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{select-set}
\alias{select-set}
\alias{ts_tree_select<-}
\alias{ts_tree_deleted}
\title{Edit parts of a tree-sitter tree}
\usage{
ts_tree_select(tree, ...) <- value
}
\arguments{
\item{tree}{A \code{ts_tree} object as returned by
\code{\link[tsitter:ts_tree_new]{ts_tree_new()}}.}

\item{...}{Selection expressions, see \code{\link[tsitter:ts_tree_select]{ts_tree_select()}}.}

\item{value}{An R expression to serialize or
\code{\link[tsitter:ts_tree_deleted]{ts_tree_deleted()}}.}
}
\value{
A \code{ts_tree} object with the selected parts updated.

\code{ts_tree_deleted()} returns a marker object to be used at the right
hand side of the \verb{ts_tree_select<-} or the double bracket replacement
functions, see examples below.
}
\description{
The \code{\link[tsitter:ts_tree_select<-]{ts_tree_select<-()}} replacement
function works similarly to the combination of
\code{\link[tsitter:ts_tree_select]{ts_tree_select()}} and
\code{\link[tsitter:ts_tree_update]{ts_tree_update()}}, but
it might be more readable.
}
\details{
The following two expressions are equivalent:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{tree <- ts_tree_update(ts_tree_select(tree, <selectors>), value)
}\if{html}{\out{</div>}}

and

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(tree, <selectors>) <- value
}\if{html}{\out{</div>}}

\code{ts_tree_deleted()} is a special marker to delete elements from a
ts_tree object with \verb{ts_tree_select<-} or the double bracket operator.
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Create a parse tree with tsjsonc -------------------------------------
tree <- tsjsonc::ts_parse_jsonc('{"a": 13, "b": [1, 2, 3], "c": "x"}')
tree

ts_tree_select(tree, "a") <- 42
ts_tree_select(tree, "b", -1) <- ts_tree_deleted()

tree
\dontshow{\}) # examplesIf}
}
\seealso{
Other ts_tree generics:
\code{\link[=[[.ts_tree]{[[.ts_tree()}},
\code{\link[=[[<-.ts_tree]{[[<-.ts_tree()}},
\code{\link[=format.ts_tree]{format.ts_tree()}},
\code{\link[=print.ts_tree]{print.ts_tree()}},
\code{\link[=ts_tree_ast]{ts_tree_ast()}},
\code{\link[=ts_tree_delete]{ts_tree_delete()}},
\code{\link[=ts_tree_dom]{ts_tree_dom()}},
\code{\link[=ts_tree_format]{ts_tree_format()}},
\code{\link[=ts_tree_insert]{ts_tree_insert()}},
\code{\link[=ts_tree_new]{ts_tree_new()}},
\code{\link[=ts_tree_query]{ts_tree_query()}},
\code{\link[=ts_tree_select]{ts_tree_select()}},
\code{\link[=ts_tree_sexpr]{ts_tree_sexpr()}},
\code{\link[=ts_tree_unserialize]{ts_tree_unserialize()}},
\code{\link[=ts_tree_update]{ts_tree_update()}},
\code{\link[=ts_tree_write]{ts_tree_write()}}
}
\concept{ts_tree generics}
