% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sitter.R
\name{syntax_tree}
\alias{syntax_tree}
\title{Show the annotated syntax tree of a file or string}
\usage{
syntax_tree(
  language,
  file = NULL,
  text = NULL,
  ranges = NULL,
  fail_on_parse_error = TRUE
)
}
\arguments{
\item{language}{Language of the file or string, a \code{ts_language} object.}

\item{file}{Path of a file. Use either \code{file} or \code{text}.}

\item{text}{String. Use either \code{file} or \code{text}, but not both.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}

\item{fail_on_parse_error}{Logical, whether to error if there are
parse errors in the document. Default is \code{TRUE}.}
}
\description{
Show the annotated syntax tree of a file or string
}
\examples{
# TODO
}
