\name{treering}
\docType{data}
\alias{treering}
\title{
  Yearly Treering Data, -6000--1979
}
\description{
  Contains normalized tree-ring widths in dimensionless units.
  Each tree ring corresponds to one year. Tree: Methuselah Walk, Pilo;
  Location: California, Gt Basin B C pine 2805M, 3726-11810; 
  Author: Donald A. Graybill, 1980.  
}
\usage{
data(treering)
}
\format{
  A univariate time series with 7981 observations. The object is of
  class \code{"ts"}.
}
\source{
  Time Series Data Library:
  \url{http://www-personal.buseco.monash.edu.au/~hyndman/TSDL/}
}
\keyword{datasets}
