% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-package.R
\docType{package}
\name{ts-package}
\alias{ts}
\alias{ts-package}
\title{ts: Tree-Sitter Parsing Tools}
\value{
Not applicable.
}
\description{
Common tree-sitter parsing tools for R. It is meant to be used by other packages that specialize in particular languages and file formats.
}
\details{
\eval{ts:::format_rd_parser_list(ts:::ts_list_parsers())}
}
\examples{
# See help("about", package = "ts")
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/ts}
  \item \url{https://r-lib.github.io/ts/}
  \item Report bugs at \url{https://github.com/r-lib/ts/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
  \item Tree-sitter authors (Tree-sitter C library) [copyright holder]
}

}
\keyword{internal}
