% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_add_sig_segments_and_labels.R
\name{TS_add_sig_segments_and_labels}
\alias{TS_add_sig_segments_and_labels}
\title{Adds test results to column charts}
\usage{
TS_add_sig_segments_and_labels(
  plot,
  compare_groups,
  gene_list_levels_plotting,
  df_stats,
  sig_distance_above_bar,
  sig_distance_between_significancies,
  sig_label_distance_from_line,
  sig_line_thickness,
  sig_label_text_size
)
}
\arguments{
\item{plot}{GGplot2 object. The output from TS_plot_column_chart()}

\item{compare_groups}{Characer vector with groups to be compared}

\item{df_stats}{Dataframe. The output from TS_compute_pairwise_fisher()}

\item{gene_list_level_plotting}{Character vector with factor levels for x_axis categories (genes)}
}
\value{
Returns a ggplot object with geom_segment() and geom_text() added
}
\description{
Helper function for TS_plot_IG_usage
Adds test results from TS_compute_pairwise_fisher() to TS_plot_column_chart() charts
}
\examples{
plot <- TS_add_sig_segments_and_labels(plot = plot,
compare_groups = compare_groups,
gene_list_levels_plotting = gene_list_levels_plotting,
df_stats = df_stats)
}
