% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_cellnumbers_per_PC_and_resolution.R
\name{TS_cellnumbers_per_PC_and_resolution}
\alias{TS_cellnumbers_per_PC_and_resolution}
\title{Calculate cellnumbers per cluster for each combination of 'number or PCs' and 'resolution'}
\usage{
TS_cellnumbers_per_PC_and_resolution(seurat_object, PCs)
}
\arguments{
\item{seurat_object}{Input seurat object}

\item{PCs}{Number of maximum PC you want to look at}
}
\description{
Calculate cellnumbers per cluster for each combination of 'number or PCs' and 'resolution'
}
