% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_compare_PC_and_resolution.R
\name{TS_compare_PC_and_resolution}
\alias{TS_compare_PC_and_resolution}
\title{Compare number of clusters for combinations of number of PCs and resolution}
\usage{
TS_compare_PC_and_resolution(seurat_object, PCs)
}
\arguments{
\item{seurat_object}{Your input seurat object}

\item{PCs}{Number of the maximum PC you want to look at}
}
\description{
Old function. Not sure of how well it functions, or what it produces.
I think this function counts the number of clusters for each combination of
\itemize{
\item 'dims = 1:i'
\item 'resolution = seq(from = 0.1, to = 2, by = 0.1)'
}
}
