% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_count_VDJ.R
\name{TS_count_VDJ}
\alias{TS_count_VDJ}
\title{Not sure if this function is desirable anymore. So use it at your own risk}
\usage{
TS_count_VDJ(vdj_file)
}
\arguments{
\item{vdj_file}{I guess it takes the input from cellranger vdj}
}
\description{
Not sure what this function does
}
