% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_count_VDJ_7.0.R
\name{TS_count_VDJ_7.0}
\alias{TS_count_VDJ_7.0}
\title{Count VDJ from Cellranger 7.0}
\usage{
TS_count_VDJ_7.0(vdj_file)
}
\arguments{
\item{vdj_file}{VDJ table which is output from the Cellranger suite.}
}
\description{
Not sure if function is usable.
}
