% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_dataframe_to_fasta.R
\name{TS_dataframe_to_fasta}
\alias{TS_dataframe_to_fasta}
\title{Converts a dataframe and saves it in Fasta format}
\usage{
TS_dataframe_to_fasta(
  dataframe,
  filename = gsub(pattern = " ", replacement = "_", x = gsub(pattern = ":", replacement =
    "-", x = paste0(Sys.time(), ".fasta")))
)
}
\arguments{
\item{dataframe}{A 2-column dataframe where:
\itemize{
\item Column 1 contains cell IDs
\item Column 2 contains the full nucleotide sequence of the contig (concatenated fwr and cdr sequences in nt-form)
}}

\item{filename}{Filename for the output FILE.fasta. Defaults to 'system date and time.fasta'}
}
\description{
Requires a dataframe with two columns:
}
