% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_DEenrichRPlot.R
\name{TS_DEenrichRPlot}
\alias{TS_DEenrichRPlot}
\title{A wrapper around the DEnrichRPlot from the 'Seurat' package}
\usage{
TS_DEenrichRPlot(
  seurat_object = get(default_seurat_object),
  ident.1 = 0,
  ident.2 = 1,
  test.use = "wilcox",
  max.genes = 500
)
}
\arguments{
\item{seurat_object}{Your 'Seurat' object. Defaults to 'default_seurat_object'}

\item{ident.1}{Name of cluster 1. Defaults to '0'}

\item{ident.2}{Name of cluster 2. Defaults to '1'}

\item{test.use}{The test you want to use for DE testing. Defaults to 'Wilcoxons ranked sum' test.}

\item{max.genes}{The maximum number of genes to use. Defaults to 500}
}
\description{
A wrapper around the DEnrichRPlot from the 'Seurat' package
}
