% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_export_cluster_ids.R
\name{TS_export_cluster_ids}
\alias{TS_export_cluster_ids}
\title{Export cell barcodes for each cluster}
\usage{
TS_export_cluster_ids(seurat_object)
}
\arguments{
\item{seurat_object}{The Seurat object you wish to retrieve the identities from. Writes to your working directory.}
}
\description{
Export cell barcodes for each cluster
}
