% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_export_named_ids.R
\name{TS_export_named_ids}
\alias{TS_export_named_ids}
\title{Export identities for each cluster, which are not numerical}
\usage{
TS_export_named_ids(seurat_object)
}
\arguments{
\item{seurat_object}{The Seurat object you wish to retrieve the identities from. Writes to working directory.}
}
\description{
Export identities for each cluster, which are not numerical
}
