% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_extract_ig_mean_median_rank_per_PC.R
\name{TS_extract_ig_mean_median_rank_per_PC}
\alias{TS_extract_ig_mean_median_rank_per_PC}
\title{Extract the mean and median values of Ig-genes from a list of PCA feature loadings}
\usage{
TS_extract_ig_mean_median_rank_per_PC(PCA_loadings)
}
\arguments{
\item{PCA_loadings}{List, the output from \link{TS_extract_ranked_genes_per_PC}}
}
\description{
Takes as input the output from \link{TS_extract_ranked_genes_per_PC}
}
