% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_extract_ig_ranks_per_PC.R
\name{TS_extract_ig_ranks_per_PC}
\alias{TS_extract_ig_ranks_per_PC}
\title{Extract the 'rank' for each of the Ig-genes}
\usage{
TS_extract_ig_ranks_per_PC(PCA_loadings)
}
\arguments{
\item{PCA_loadings}{The output from \link{TS_extract_ranked_genes_per_PC}}
}
\description{
Takes as input the output from \link{TS_extract_ranked_genes_per_PC}
}
