% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_extract_ranked_genes_per_PC.R
\name{TS_extract_ranked_genes_per_PC}
\alias{TS_extract_ranked_genes_per_PC}
\title{Extract an ordered list of feature importance for each of the first 50 principal components}
\usage{
TS_extract_ranked_genes_per_PC(seurat_object)
}
\arguments{
\item{seurat_object}{The Seurat object you wish to extract PCA feature loadings from}
}
\description{
Extract an ordered list of feature importance for each of the first 50 principal components
}
