% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_featureplot_to_pdf.R
\name{TS_featureplot_to_pdf}
\alias{TS_featureplot_to_pdf}
\title{Export featureplot as pdf to your working directory}
\usage{
TS_featureplot_to_pdf(
  seurat_object,
  file_prefix,
  features,
  pt.size = 0.5,
  order = TRUE,
  cols = c("lightgray", "#FEFDE2", "#FEEAC7", "#FDD7AC", "#FDC591", "#FDB276", "#FC9F5B",
    "#FC8C40", "#FC8C40", "#EB753C", "#DA5D38", "#C94634", "#B72F30", "#A6172C",
    "#950028"),
  reduction = "umap"
)
}
\arguments{
\item{seurat_object}{The Seurat object you wish to plot}

\item{file_prefix}{Prefix for the saved file, e.g. sample name}

\item{features}{Vector of features to plot}

\item{pt.size}{Dot size, default = 0.5}

\item{order}{Whether to plot positive cells last, e.g. on top, default = TRUE}

\item{cols}{Vector of 2 colors to use (low->high)}

\item{reduction}{Dimensionality reduction technique to use for plotting, default = "umap"}
}
\description{
Export featureplot as pdf to your working directory
}
