% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_featureplot_to_pdf_A4.R
\name{TS_featureplot_to_pdf_A4}
\alias{TS_featureplot_to_pdf_A4}
\title{Export featureplot to pdf in A4 format}
\usage{
TS_featureplot_to_pdf_A4(
  seurat_object,
  file_prefix,
  features,
  pt.size = 1.5,
  reduction = "umap",
  order = TRUE,
  cols = c("lightgray", "#FEFDE2", "#FEEAC7", "#FDD7AC", "#FDC591", "#FDB276", "#FC9F5B",
    "#FC8C40", "#FC8C40", "#EB753C", "#DA5D38", "#C94634", "#B72F30", "#A6172C",
    "#950028")
)
}
\arguments{
\item{seurat_object}{The Seurat object you wish to plot}

\item{file_prefix}{Prefix for the saved file, e.g. sample name}

\item{features}{Vector of features to plot}

\item{pt.size}{Dot size, default = 1.5}

\item{reduction}{Dimensionality reduction technique to use for plotting, default = "umap"}

\item{order}{Whether to plot positive cells last, e.g. on top, default = 1.5}

\item{cols}{Vector of 2 colors to use (low->high), default = c("lightgray", "red")}
}
\description{
Export featureplot to pdf in A4 format
}
