% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_featureplot_to_pdf_no_leg_no_axes.R
\name{TS_featureplot_to_pdf_no_leg_no_axes}
\alias{TS_featureplot_to_pdf_no_leg_no_axes}
\title{Export featureplot to pdf without legend and axes}
\usage{
TS_featureplot_to_pdf_no_leg_no_axes(
  seurat_object,
  file_prefix,
  features,
  pt.size = 0.5,
  order = TRUE,
  cols = c("lightgray", "#FEFDE2", "#FEEAC7", "#FDD7AC", "#FDC591", "#FDB276", "#FC9F5B",
    "#FC8C40", "#FC8C40", "#EB753C", "#DA5D38", "#C94634", "#B72F30", "#A6172C",
    "#950028"),
  reduction = "umap",
  width = 4.4,
  height = 3.78
)
}
\arguments{
\item{seurat_object}{The Seurat object you wish to plot}

\item{file_prefix}{Prefix for the saved file, e.g. sample name}

\item{features}{Vector of features to plot}

\item{pt.size}{Dot size, default = 0.5}

\item{order}{Whether to plot positive cells last, e.g. on top, default = TRUE}

\item{cols}{Vector of 2 colors to use (low->high), default = c("lightgray", "red")}

\item{reduction}{Dimensionality reduction technique to use for plotting, default = "umap"}

\item{width}{Width (inches) of the resulting plot.}

\item{height}{Height (inches) of the resulting plot.}
}
\description{
Export featureplot to pdf without legend and axes
}
