% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_featureplots.R
\name{TS_featureplots}
\alias{TS_featureplots}
\title{Output one or multiple featureplots}
\usage{
TS_featureplots(
  features,
  seurat_object = get(default_seurat_object),
  wrap_plots = FALSE,
  wrap_columns = NULL,
  cols = c("lightgray", "#FEFDE2", "#FEEAC7", "#FDD7AC", "#FDC591", "#FDB276", "#FC9F5B",
    "#FC8C40", "#FC8C40", "#EB753C", "#DA5D38", "#C94634", "#B72F30", "#A6172C",
    "#950028"),
  reduction = "umap",
  pt.size = 1,
  order = T
)
}
\arguments{
\item{features}{Vector of features to plot.}

\item{seurat_object}{The Seurat object you wish to plot.}

\item{wrap_plots}{Whether to wrap plots into one by patchwork::wrap_plots().}

\item{wrap_columns}{If \link{wrap_plots = T}, how many columns do you want for the final plot.}

\item{cols}{Vector of colors to use (low->high), defaults to a 15-step gradient from "lightgray" to "#950028"}

\item{reduction}{Dimensionality reduction technique to use for plotting, default = "umap".}

\item{pt.size}{Size of each dot, default = 1.}

\item{order}{Whether to plot positive cells last, e.g. on top, default = TRUE.}
}
\description{
Output one or multiple featureplots
}
