% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_filter_marker_list.R
\name{TS_filter_marker_list}
\alias{TS_filter_marker_list}
\title{Filter list of markers from FindMarkers function, needs to be output into new df}
\usage{
TS_filter_marker_list(df, markers_to_remove)
}
\arguments{
\item{df}{The dataframe you wish to filter}

\item{markers_to_remove}{Vector of markers to remove}
}
\description{
Filter list of markers from FindMarkers function, needs to be output into new df
}
