% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_FindConservedMarkers.R
\name{TS_FindConservedMarkers}
\alias{TS_FindConservedMarkers}
\title{Find conserved markers within two groups}
\usage{
TS_FindConservedMarkers(
  seurat_object,
  ident.1,
  ident.2,
  grouping_variable_name = "tested_population",
  only_pos_markers = T
)
}
\arguments{
\item{seurat_object}{Your Seurat object}

\item{ident.1}{Name for cluster 1}

\item{ident.2}{Name for cluster 2}

\item{grouping_variable_name}{Name of the combined clusters, can be anything. Defaults to 'tested_population'.}

\item{only_pos_markers}{Whether to only highlight positively upregulated markers between the grouped population and the rest. Defaults to 'TRUE'.}
}
\description{
Not sure this function does what I want it to...
}
