% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_IGHV_usage_prop_with_stats.R
\name{TS_IGHV_usage_prop_with_stats}
\alias{TS_IGHV_usage_prop_with_stats}
\title{Plots IGHV usage between groups together with Fishers Exact Test}
\usage{
TS_IGHV_usage_prop_with_stats(
  input_file,
  sample1,
  sample2,
  export_pdf = F,
  return_only_data = F,
  ylim_min = 0,
  ylim_max = 15,
  p.adj.method = "hochberg"
)
}
\arguments{
\item{input_file}{Takes as input a dataframe from 10X Genomixs Cellranger VDJ pipeline that has been through QC with added sample/grouping information in the column "sample"}

\item{sample1}{Grouping variable value for sample 1}

\item{sample2}{Grouping variable value for sample 2}

\item{export_pdf}{Whether to automatically export the plot as PDF in your working directory. Defaults to 'FALSE'}

\item{return_only_data}{Whether to only return the the calculated frequencies and the test matrices. 'FALSE' returns the plot in standard out. Defaults to 'FALSE'}

\item{ylim_min}{Defaults to '0'}

\item{ylim_max}{Defaults to '15'. Values above this will be removed.}

\item{p.adj.method}{Method for adjust for multiple tests. Defaults to 'Benjamini-Hochberg'. Alternative is 'FDR'.}
}
\description{
Compare the IGHV usage between two groups and perform Fishers Exact Test to test usage.
This also means that you can only input two groups, due to limiations to the statistical test used. Future update?
}
