% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plot_column_pattern_chart.R
\name{TS_plot_column_pattern_chart}
\alias{TS_plot_column_pattern_chart}
\title{Create column chart with pattern with ggplot2}
\usage{
TS_plot_column_pattern_chart(
  data,
  x,
  y,
  fill,
  fill_colours,
  name = "",
  ggplot_theme,
  ylim_max,
  border_colour,
  plot_title,
  plot_subtitle,
  pattern,
  pattern_colour = "black",
  pattern_fill = "black",
  pattern_density = 0.1,
  plot_n_sequences
)
}
\arguments{
\item{x}{Variable to plot on x-axis.}

\item{y}{Variable to plot on y-axis.}

\item{fill}{Variable to fill on (group variable).}

\item{fill_colours}{Vector. Fill colours to use.}

\item{name}{String. Legend title. Defaults to "Sample"}

\item{ggplot_theme}{Optional. A specified ggplot_theme to use.}

\item{plot_data}{A dataframe containing the long format data which you want to plot with precomputed proportions.}

\item{ylim}{Optional, numeric. Use to specify max y_lim value.}
}
\value{
A tibble with columns: gene, group1, group2, p_value, p_adj
}
\description{
Create column chart with pattern with ggplot2
}
