% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plot_counts_boxplot.R
\name{TS_plot_counts_boxplot}
\alias{TS_plot_counts_boxplot}
\title{Plot counts for a gene in a Seurat object as boxplot}
\usage{
TS_plot_counts_boxplot(
  seurat_object = get(default_seurat_object),
  slot = "counts",
  gene
)
}
\arguments{
\item{seurat_object}{The Seurat object you want to analyse}

\item{slot}{Slot to plot. Defaults to "counts"}

\item{gene}{Gene to plot}
}
\description{
Plot counts for a gene in a Seurat object as boxplot
}
