% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plot_enrichR.R
\name{TS_plot_enrichR}
\alias{TS_plot_enrichR}
\title{Plot the output from \link{TS_run_enrichR}}
\usage{
TS_plot_enrichR(input, sort_on = "neg_log10", n = 10)
}
\arguments{
\item{input}{The output list from \link{TS_run_enrichR}}

\item{sort_on}{What to sort top n terms on. Defaults to 'neg_log10', can also be 'combined_score'}

\item{n}{Number of terms to plot for each database. Defaults to '10'.}
}
\description{
A short description...
}
