% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_plot_IG_usage.R
\name{TS_plot_IG_usage}
\alias{TS_plot_IG_usage}
\title{Plots Ig usage between two groups using Fisher's Exact Test}
\usage{
TS_plot_IG_usage(
  input_file,
  sample1,
  sample2,
  compare_what = "IGHV",
  grouping_variable = "named_clusters",
  export_pdf = F,
  return_only_data = F,
  ylim_max = NULL,
  p.adj.method = "hochberg",
  fill_colours = c("gray80", "gray50"),
  border_colour = "black",
  ggplot_theme = ggplot2::theme_classic(),
  plot_width = 15,
  plot_height = 5,
  remove_genes_present_in_one_dataset = F,
  plot_zeroes = T,
  file_type = "cellranger"
)
}
\arguments{
\item{input_file}{A dataframe (e.g. from 10X Genomics Cellranger VDJ output) that includes a grouping column.}

\item{sample1}{Grouping variable value for sample 1}

\item{sample2}{Grouping variable value for sample 2}

\item{compare_what}{Genes to compare (e.g. "IGHV", "IGKV"). Defaults to "IGHV".}

\item{grouping_variable}{Column used to group data. Defaults to "named_clusters"}

\item{export_pdf}{Logical. If TRUE, exports the plot to a PDF. Defaults to FALSE}

\item{return_only_data}{Logical. If TRUE, returns only the calculated frequencies and the test matrices. Defaults to FALSE.}

\item{ylim_max}{Numeric. Optional upper limit for y-axis. ggplot2 will adjust the y-axis scale to fit the data.}

\item{p.adj.method}{Method for adjust for multiple tests. Defaults to 'Hochberg'. Alternative is 'FDR'.}

\item{fill_colours}{A character vector of fill colours for the two groups. Defaults to c("grey80", "grey50").}

\item{border_colour}{Colour of bar borders. Defaults to "black".}

\item{ggplot_theme}{Optional ggplot2 theme to apply. Defaults to "theme_classic".}

\item{plot_width}{Numeric. Width of exported plot in inches. Used if 'export_pdf = TRUE'. Defaults to 15.}

\item{plot_height}{Numeric. Height of exported plot in inches. Used if 'export_pdf = TRUE'. Defaults to 5.}

\item{remove_genes_present_in_one_dataset}{Logical. If TRUE, removes genes not shared by both groups. Defaults to FALSE}

\item{plot_zeroes}{Logical. If FALSE, genes with 0 counts in one group are not plotted, resulting in a wider bar for the other group.}

\item{file_type}{Set type of file/annotation input. Defaults to "cellranger". Options are "IMGT" and "AIRR".}
}
\value{
Returns a ggplot object, or a data.frame containing counts and test results if ' return_only_data = TRUE'.
}
\description{
Plot the Ig usage from two groups and perform Fisher's Exact Test to find significant differences.
Maximum number of groups = 2
}
\examples{
# TS_plot_IG_usage(input_file = VDJ_data, sample1 = "group_A", sample2 = "group_B", compare_what = "IGHV")
}
