% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_project_scRNAseq_dataset.R
\name{TS_project_scRNAseq_dataset}
\alias{TS_project_scRNAseq_dataset}
\title{Projects one Seurat object onto another}
\usage{
TS_project_scRNAseq_dataset(
  reference_object,
  query_object,
  ref_name,
  query_name
)
}
\arguments{
\item{reference_object}{A Seurat object, will be used as reference.}

\item{query_object}{A Seurat object, will be used projected onto your reference object.}

\item{ref_name}{Name of reference object. Used for resulting plots.}

\item{query_name}{Name of query object. Used for resulting plots.}
}
\description{
A work in progress...
Will output resulting DimPlots with data from \link[Seurat:MapQuery]{Seurat::MapQuery}
}
