% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_project_scRNAseq_dataset_export_query.R
\name{TS_project_scRNAseq_dataset_export_query}
\alias{TS_project_scRNAseq_dataset_export_query}
\title{Projects one Seurat object onto another and exports resulting dataframe}
\usage{
TS_project_scRNAseq_dataset_export_query(
  reference_object,
  query_object,
  dims = 1:20,
  reference_object_assay = "integrated",
  query_object_assay = "integrated",
  plot_3D = F
)
}
\arguments{
\item{reference_object}{A Seurat object, will be used as reference.}

\item{query_object}{A Seurat object, will be used projected onto your reference object.}

\item{dims}{Number of Principal Components to use for the UMAP. Defaults to '1:20'}

\item{reference_object_assay}{Assay to use for reference Seurat object. Defaults to 'integrated'}

\item{query_object_assay}{Assay to use for query Seurat object. Defaults to 'integrated'}

\item{plot_3D}{Whether to calculate 3 dimensions for the UMAP model. Defaults to 'FALSE'}
}
\description{
A work in progress...
Will output resulting dataframe from \link[Seurat:MapQuery]{Seurat::MapQuery}
}
