% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_run_enrichR.R
\name{TS_run_enrichR}
\alias{TS_run_enrichR}
\title{Wrapper around the enrichR package.}
\usage{
TS_run_enrichR(
  gene_list,
  plot_enrichr = F,
  database_to_plot = "GO_Biological_Process_2023",
  databases = c("GO_Biological_Process_2023", "GO_Molecular_Function_2023",
    "GO_Cellular_Component_2023")
)
}
\arguments{
\item{gene_list}{Character vector containing genes to test}

\item{plot_enrichr}{Whether to plot the output using the enrichR function \link{plotEnrich}. Defaults to 'FALSE'}

\item{database_to_plot}{What database to plot the output from. Defaults to '"GO_Biological_Process_2023"'}

\item{databases}{What databases to test against. Defaults to 'c("GO_Biological_Process_2023", "GO_Molecular_Function_2023", "GO_Cellular_Component_2023")'}
}
\description{
Performs enrichment analyses for a vector of genes
Depends on
-enrichR (and that their servers are up, heh)
\itemize{
\item A live Internet connection
\item magrittr
\item dplyr
Returns a list of the results of each database tested for.
}
}
