% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-select.R
\name{ts_select}
\alias{ts_select}
\title{Select parts of a tree-sitter tree}
\usage{
ts_select(tokens, ...)
}
\arguments{
\item{tokens}{A \code{ts_tokens} object as returned by \code{\link[=ts_parse]{ts_parse()}}.}

\item{...}{Selection expressions, see details.}
}
\description{
TODO: include links to methods here dynamically?
}
