% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_seurat_object_stats.R
\name{TS_seurat_object_stats}
\alias{TS_seurat_object_stats}
\title{Prints stats about your Seurat object}
\usage{
TS_seurat_object_stats(seurat_object)
}
\arguments{
\item{seurat_object}{Your Seurat object}
}
\description{
Prints stats about your Seurat object
}
