% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-sepxr.R
\name{ts_sexpr}
\alias{ts_sexpr}
\title{Show the syntax tree structure of a file or string}
\usage{
ts_sexpr(tokens)
}
\arguments{
\item{tokens}{A \code{ts_tokens} object as returned by \code{\link[=ts_parse]{ts_parse()}}.}
}
\description{
Show the syntax tree structure of a file or string
}
\examples{
# TODO
}
