% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_svg_from_plot.R
\name{TS_svg_from_plot}
\alias{TS_svg_from_plot}
\title{Export any plot as svg}
\usage{
TS_svg_from_plot(plot, filename)
}
\arguments{
\item{plot}{Needs a ggplot as input}

\item{filename}{Filename for the resulting plot, needs to be within "".}
}
\description{
Export any plot as svg
}
