% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-syntax-tree.R
\name{ts_syntax_tree}
\alias{ts_syntax_tree}
\title{Show the annotated syntax tree of a file or string}
\usage{
ts_syntax_tree(tokens)
}
\arguments{
\item{tokens}{A \code{ts_tokens} object as returned by \code{\link[=ts_parse]{ts_parse()}}.}
}
\description{
Show the annotated syntax tree of a file or string
}
\examples{
# TODO
}
