% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket-ts-tokens.R
\name{ts_tokens-brackets}
\alias{ts_tokens-brackets}
\alias{[.ts_tokens}
\title{Convert ts_tokens object to a data frame}
\usage{
\method{[}{ts_tokens}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{ts_tokens object.}

\item{i, j}{indices.}

\item{drop}{Ignored.}
}
\value{
A data frame with columns: TODO.
}
\description{
Create a data frame for the syntax tree of a JSON document, by indexing
a ts_tokens object with single brackets. This is occasionally useful for
exploration and debugging.
}
\examples{
# TODO
}
\seealso{
TODO
}
