% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-read.R
\name{ts_tree_read}
\alias{ts_tree_read}
\title{Read (parse) a file or a string}
\usage{
ts_tree_read(
  language,
  file = NULL,
  text = NULL,
  ranges = NULL,
  fail_on_parse_error = TRUE
)
}
\arguments{
\item{language}{Language of the file or string, a \code{ts_language} object.}

\item{file}{Path of a file. Use either \code{file} or \code{text}.}

\item{text}{String. Use either \code{file} or \code{text}, but not both.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}

\item{fail_on_parse_error}{Logical, whether to error if there are
parse errors in the document. Default is \code{TRUE}.}
}
\value{
A data frame with one row per token, and columns:
\itemize{
\item \code{id}: integer, the id of the token.
\item \code{parent}: integer, the id of the parent token. The root token has
parent \code{NA}
\item \code{field_name}: character, the field name of the token in its parent.
\item \code{type}: character, the type of the token.
\item \code{code}: character, the actual code of the token.
\item \code{start_byte}, \code{end_byte}: integer, the byte positions of the token
in the input.
\item \code{start_row}, \code{start_column}, \code{end_row}, \code{end_column}: integer, the
position of the token in the input.
\item \code{is_missing}: logical, whether the token is a missing token added by
the parser to recover from errors.
\item \code{has_error}: logical, whether the token has a parse error.
\item \code{children}: list of integer vectors, the ids of the children tokens.
}
}
\description{
Read (parse) a file or a string
}
\examples{
# TODO
}
