% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select_query}
\alias{ts_tree_select_query}
\title{TODO: move this into ts_select}
\usage{
ts_tree_select_query(tree, query)
}
\arguments{
\item{tree}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_new]{ts_tree_new()}}.}

\item{query}{String, a tree-sitter query.}
}
\value{
TODO
}
\description{
TODO: move this into ts_select
}
