% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select_refine}
\alias{ts_tree_select_refine}
\title{TODO: move this into ts_select}
\usage{
ts_tree_select_refine(tree, ...)
}
\arguments{
\item{tree}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_read]{ts_tree_read()}}.}

\item{...}{Selection expressions.}
}
\description{
TODO: move this into ts_select
}
