% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_selection}
\alias{ts_tree_selection}
\alias{ts_tree_selected_nodes}
\title{Helper functions for tree-sitter tree selections}
\usage{
ts_tree_selection(tree, default = TRUE)

ts_tree_selected_nodes(tree, default = TRUE)
}
\arguments{
\item{tree}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_read]{ts_tree_read()}}.}

\item{default}{Logical, whether to return the default selection if there
is no explicit selection, or \code{NULL}.}
}
\value{
\code{ts_tree_selection()} returns a list of selection records.

\code{ts_tree_selected_nodes()} returns the ids of the currently
selected nodes.
}
\description{
It is unlikely that you will need to use these functions directly, except
when implementing a new language for the ts package.
}
\details{
\code{ts_tree_selection()} returns the current selection, as a list
of selectors.

\code{ts_tree_selected_nodes()} returns the ids of the currently
selected nodes.
}
