% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_selector_ids}
\alias{ts_tree_selector_ids}
\title{TODO: move this into ts_select}
\usage{
ts_tree_selector_ids(ids)
}
\arguments{
\item{ids}{Integer vector, the ids of the nodes to select.}
}
\description{
TODO: move this into ts_select
}
