% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_visdimloadings.R
\name{TS_visdimloadings}
\alias{TS_visdimloadings}
\title{Plot gene contribution to each principal component}
\usage{
TS_visdimloadings(seurat_object = get(default_seurat_object), dims = 1:10)
}
\arguments{
\item{seurat_object}{The Seurat object you want to plot}

\item{dims}{What principal components to plot, default = 10}
}
\description{
Plot gene contribution to each principal component
}
