% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-package.R
\name{about}
\alias{about}
\title{About tsitter}
\value{
Not applicable.
}
\description{
tsitter is a common interface to \href{https://tree-sitter.github.io/tree-sitter/}{tree-sitter} parsers, implemented in
other R packages. It has a common API to
\itemize{
\item query,
\item edit,
\item format, and
\item unserialize
}

tree-sitter parse trees.
}
\details{
In this document I show examples with the \code{tsjsonc} package.
\subsection{Create a tree-sitter tree}{

Create a ts_tree (ts_tree_jsonc) object from a string:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{txt <- r"(
// this is a comment
\{
  "a": \{
    "a1": [1, 2, 3],
    // comment
    "a2": "string"
  \},
  "b": [
    \{
      "b11": true,
      "b12": false
    \},
    \{
      "b21": false,
      "b22": false
    \}
  ]
\}
)"

json <- tsjsonc::ts_parse_jsonc(txt)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#>                                                                                 
</pre></div>
}}


Pretty print a ts_tree object:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{json
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># jsonc (19 lines)</span>                                                              
#> <span style="color: #6F7783;"> 1 | </span>                                                                           
#> <span style="color: #6F7783;"> 2 | </span>// this is a comment                                                       
#> <span style="color: #6F7783;"> 3 | </span>\{                                                                          
#> <span style="color: #6F7783;"> 4 | </span>  "a": \{                                                                   
#> <span style="color: #6F7783;"> 5 | </span>    "a1": [1, 2, 3],                                                       
#> <span style="color: #6F7783;"> 6 | </span>    // comment                                                             
#> <span style="color: #6F7783;"> 7 | </span>    "a2": "string"                                                         
#> <span style="color: #6F7783;"> 8 | </span>  \},                                                                       
#> <span style="color: #6F7783;"> 9 | </span>  "b": [                                                                   
#> <span style="color: #6F7783;">10 | </span>    \{                                                                      
#> <span style="color: #6F7783;">ℹ 9 more lines</span>                                                                  
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}

}

\subsection{Select nodes of a tree}{

Selecting nodes is the basis of editing and querying tree-sitter trees.

Select element by objects key:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(json, "a")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_select(json, "a") :                                            
#>   could not find function "ts_tree_select"                                      
</pre></div>
}}


Select element inside element:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(json, "a", "a1")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_select(json, "a", "a1") :                                      
#>   could not find function "ts_tree_select"                                      
</pre></div>
}}


Select element(s) of an array:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(json, "a", "a1", 1:2)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_select(json, "a", "a1", 1:2) :                                 
#>   could not find function "ts_tree_select"                                      
</pre></div>
}}


Select multiple keys from an object:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(json, "a", c("a1", "a2"))
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_select(json, "a", c("a1", "a2")) :                             
#>   could not find function "ts_tree_select"                                      
</pre></div>
}}


Select nodes that match a tree-sitter query:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{json |> ts_tree_select(query = "((pair value: (false) @val))")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_select(json, query = "((pair value: (false) @val))") :         
#>   could not find function "ts_tree_select"                                      
</pre></div>
}}

}

\subsection{Delete elements}{

Delete selected elements:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(json, "a", "a1") |> ts_tree_delete()
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_delete(ts_tree_select(json, "a", "a1")) :                      
#>   could not find function "ts_tree_delete"                                      
</pre></div>
}}

}

\subsection{Insert elements}{

Insert element into an array:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(json, "a", "a1") |> ts_tree_insert(at = 2, "new")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_insert(ts_tree_select(json, "a", "a1"), at = 2, "new") :       
#>   could not find function "ts_tree_insert"                                      
</pre></div>
}}


Inserting into an array reformats the array.

Insert element into an object, at the specified key:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(json, "a") |>
  ts_tree_insert(key = "a0", at = 0, list("new", "element"))
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_insert(ts_tree_select(json, "a"), key = "a0", at = 0,  :       
#>   could not find function "ts_tree_insert"                                      
</pre></div>
}}

}

\subsection{Update elements}{

Update existing element:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(json, "a", c("a1", "a2")) |> ts_tree_update("new value")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#>                                                                                 
#>   could not find function "ts_tree_update"                                      
</pre></div>
}}


Inserts the element if some parents are missing:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{json <- ts_parse_jsonc(text = "\{ \\"a\\": \{ \\"b\\": true \} \}")
json
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_parse_jsonc(text = "\{ \\"a\\": \{ \\"b\\": true \} \}") :                  
#>   could not find function "ts_parse_jsonc"                                      
#> <span style="color: #6F7783;"># jsonc (19 lines)</span>                                                              
#> <span style="color: #6F7783;"> 1 | </span>                                                                           
#> <span style="color: #6F7783;"> 2 | </span>// this is a comment                                                       
#> <span style="color: #6F7783;"> 3 | </span>\{                                                                          
#> <span style="color: #6F7783;"> 4 | </span>  "a": \{                                                                   
#> <span style="color: #6F7783;"> 5 | </span>    "a1": [1, 2, 3],                                                       
#> <span style="color: #6F7783;"> 6 | </span>    // comment                                                             
#> <span style="color: #6F7783;"> 7 | </span>    "a2": "string"                                                         
#> <span style="color: #6F7783;"> 8 | </span>  \},                                                                       
#> <span style="color: #6F7783;"> 9 | </span>  "b": [                                                                   
#> <span style="color: #6F7783;">10 | </span>    \{                                                                      
#> <span style="color: #6F7783;">ℹ 9 more lines</span>                                                                  
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}


\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(json, "a", "x", "y") |> ts_tree_update(list(1,2,3))
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_update(ts_tree_select(json, "a", "x", "y"), list(1, 2,  :      
#>   could not find function "ts_tree_update"                                      
</pre></div>
}}

}

\subsection{Write out a document}{

Use \code{stdout()} to write it to the screen instread of a file:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{json |> ts_tree_write(stdout())
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_write(json, stdout()) :                                        
#>   could not find function "ts_tree_write"                                       
</pre></div>
}}

}

\subsection{Formatting}{

Format the whole document:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{json |> ts_tree_format()
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_format(json) : could not find function "ts_tree_format"        
</pre></div>
}}


Format part of the document:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{json |> ts_tree_select("a") |>
  ts_tree_format(options = list(format = "compact"))
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> act")) :                                                                        
#>   could not find function "ts_tree_format"                                      
</pre></div>
}}

}

\subsection{Unserializing}{

Unserialize a whole document:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{json |> ts_tree_unserialize()
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_unserialize(json) :                                            
#>   could not find function "ts_tree_unserialize"                                 
</pre></div>
}}


Note that \code{ts_tree_unserialize()} always returns a list, the first element
of the list is the unserialized document.

Unserialize part(s) of the document:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{json |> ts_tree_select("b") |> ts_tree_unserialize()
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_unserialize(ts_tree_select(json, "b")) :                       
#>   could not find function "ts_tree_unserialize"                                 
</pre></div>
}}


Again, \code{ts_tree_unserialize()} returns a list, with one element for each
selected node.
}

\subsection{Exploring a tree-sitter tree}{

It is often useful to explore the structure of a (JSONC) tree-sitter tree,
to help writing the right selection or tree-sitter queries.

Print the annotated syntax tree:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_ast(json)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_ast(json) : could not find function "ts_tree_ast"              
</pre></div>
}}


Print the document object model:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_dom(json)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_dom(json) : could not find function "ts_tree_dom"              
</pre></div>
}}


Print the structural summary of a tree:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_sexpr(json)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> Error in ts_tree_sexpr(json) : could not find function "ts_tree_sexpr"          
#>                                                                                 
#>                                                                                 
#>                                                                                 
#>                                                                                 
#>                                                                                 
#>                                                                                 
</pre></div>
}}

}
}
\examples{
# See above please.

}
