% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{[[<-.ts_tree}
\alias{[[<-.ts_tree}
\title{Edit parts of a tree-sitter tree}
\usage{
\method{[[}{ts_tree}(x, i) <- value
}
\arguments{
\item{x}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_new]{ts_tree_new()}}.}

\item{i}{A list with selection expressions, see details.}
}
\description{
TODO
}
