% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{[[.ts_tree}
\alias{[[.ts_tree}
\title{Unserialize selected parts of a tree-sitter tree}
\usage{
\method{[[}{ts_tree}(x, i, ...)
}
\arguments{
\item{x}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_new]{ts_tree_new()}}.}

\item{i}{Selection expressions, see details in \code{\link[=ts_tree_select]{ts_tree_select()}}.}

\item{...}{Passed to \code{\link[=ts_tree_select]{ts_tree_select()}}.}
}
\description{
TODO
}
