% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{[[.ts_tree}
\alias{[[.ts_tree}
\title{Unserialize parts of a tree-sitter tree}
\usage{
\method{[[}{ts_tree}(x, i, ...)
}
\arguments{
\item{x}{\eval{ts:::doc_insert("ts::ts_tree_double_bracket_param_x")}}

\item{i}{\eval{ts:::doc_insert("ts::ts_tree_double_bracket_param_i")}}

\item{...}{\eval{ts:::doc_insert("ts::ts_tree_double_bracket_param_dots")}}
}
\value{
\eval{ts:::doc_insert("ts::ts_tree_double_bracket_return")}
}
\description{
\eval{ts:::doc_insert("ts_tree_double_bracket_description")}
}
\details{
\eval{ts:::doc_insert("ts::ts_tree_double_bracket_details")}
\eval{ts:::doc_extra()}
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Create a parse tree with tsjsonc -------------------------------------
tree <- tsjsonc::ts_parse_jsonc('{"a": 13, "b": [1, 2, 3], "c": "x"}')

tree

tree[[list("a")]]

# Last two elements of "b"
tree[[list("b", -(1:2))]]
\dontshow{\}) # examplesIf}
}
\seealso{
Other ts_tree generics: 
\code{\link{[[<-.ts_tree}()},
\code{\link{format.ts_tree}()},
\code{\link{print.ts_tree}()},
\code{\link{select-set}},
\code{\link{ts_tree_ast}()},
\code{\link{ts_tree_delete}()},
\code{\link{ts_tree_dom}()},
\code{\link{ts_tree_format}()},
\code{\link{ts_tree_insert}()},
\code{\link{ts_tree_new}()},
\code{\link{ts_tree_query}()},
\code{\link{ts_tree_select}()},
\code{\link{ts_tree_sexpr}()},
\code{\link{ts_tree_unserialize}()},
\code{\link{ts_tree_update}()},
\code{\link{ts_tree_write}()}

Other serialization functions: 
\code{\link{ts_tree_unserialize}()}
}
\concept{serialization functions}
\concept{ts_tree generics}
