% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.ts_tree}
\alias{format.ts_tree}
\title{Format tree-sitter trees}
\usage{
\method{format}{ts_tree}(x, n = 10, ...)
}
\arguments{
\item{x}{ts_tree  object.}

\item{n}{Number of lines, or number of selections to print.}

\item{...}{Ignored.}
}
\value{
Character vector.
}
\description{
Format a ts_tree object for printing.
}
\details{
This is the engine of \code{\link[=print.ts_tree]{print.ts_tree()}}, possibly useful to obtain a
printed representation without doing the actual printing.
}
\examples{
# TODO
}
