% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.ts_tree}
\alias{format.ts_tree}
\title{Format tree-sitter trees}
\usage{
\method{format}{ts_tree}(x, n = 10, ...)
}
\arguments{
\item{x}{\eval{ts:::doc_insert("ts::format_param_x")}}

\item{n}{\eval{ts:::doc_insert("ts::format_param_n")}}

\item{...}{\eval{ts:::doc_insert("ts::format_param_dots")}}
}
\value{
\eval{ts:::doc_insert("ts::format_return")}
}
\description{
\eval{ts:::doc_insert("ts::format_description")}
}
\details{
\eval{ts:::doc_insert("ts::format_details")}
\eval{ts:::doc_extra()}
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Create a parse tree with tsjsonc -------------------------------------
json <- tsjsonc::ts_parse_jsonc(
  '{ "a": 1, "b": [10, 20, 30], "c": { "c1": true, "c2": 100 } }'
)
format(json)
\dontshow{\}) # examplesIf}
}
\seealso{
Other \code{ts_tree} generics: 
\code{\link{[[.ts_tree}()},
\code{\link{[[<-.ts_tree}()},
\code{\link{print.ts_tree}()},
\code{\link{select-set}},
\code{\link{ts_tree_ast}()},
\code{\link{ts_tree_delete}()},
\code{\link{ts_tree_dom}()},
\code{\link{ts_tree_format}()},
\code{\link{ts_tree_insert}()},
\code{\link{ts_tree_new}()},
\code{\link{ts_tree_query}()},
\code{\link{ts_tree_select}()},
\code{\link{ts_tree_sexpr}()},
\code{\link{ts_tree_unserialize}()},
\code{\link{ts_tree_update}()},
\code{\link{ts_tree_write}()}
}
\concept{\code{ts_tree} generics}
