% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ts_tree}
\alias{print.ts_tree}
\title{Print a tree-sitter tree}
\usage{
\method{print}{ts_tree}(x, n = 10, ...)
}
\arguments{
\item{x}{ts_tree object.}

\item{n}{Number of lines, or number of selections to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Calls \code{\link[=format.ts_tree]{format.ts_tree()}} to format the ts_tree object, writes the
formatted object to the standard output, and returns the original object
invisibly.
}
\examples{
# TODO
}
