\name{sales}
\alias{sales}
\title{
  Sales Data with Leading Indicator.
}
\usage{
data (sales)
}
\format{
  The sales time series \code{sales} and leading indicator
  \code{sales.lead} each contain 150 observations.
  The objects are of class \code{"ts"}.
}
\description{
  The data is given in G. E. P. Box and G. M. Jenkins (1976): \emph{Time Series
  Analysis, Forecasting and Control}, Holden-Day, San
  Francisco, p. 537.

  See also P. J. Brockwell and R. A. Davis (1991): \emph{Time Series:
  Theory and Methods}, 2nd Edition, Springer Verlag, NY,
  pp. 414.
}
\source{
  Time Series Data Library
  http://www-personal.buseco.monash.edu.au/~hyndman/TSDL/
}
\keyword{datasets}


