% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{select-set}
\alias{select-set}
\alias{ts_tree_select<-}
\alias{[[<-.ts_tree}
\alias{ts_tree_deleted}
\title{TODO}
\usage{
ts_tree_select(tree, ...) <- value

\method{[[}{ts_tree}(x, i) <- value
}
\arguments{
\item{tree, x}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_new]{ts_tree_new()}}.}

\item{...}{Selection expressions, see details.}

\item{value}{An R expression to serialize or \code{ts_tree_deleted()}.}

\item{i}{A list with selection expressions, see details.}
}
\value{
The modified \code{ts_tree} object.

\code{ts_tree_deleted()} returns a marker object to be used at the right
hand side of the \verb{ts_tree_select<-} or the double bracket replacement
functions, see examples below.
}
\description{
TODO

TODO
}
\details{
\code{ts_tree_deleted()} is a special marker to delete elements from a tree-sitter
ts_tree object with \verb{ts_tree_select<-} or the double bracket operator.
}
\examples{
# TODO
# TODO
}
