\name{treering}
\alias{treering}
\title{
  Yearly Treering Data, -6000-1979.
}
\usage{
data (treering)
}
\format{
    A univariate time series with 7981 observations. The object is of
    class \code{"ts"}.
}
\description{
  Contains normalized tree-ring widths in dimensionless units.
  Each tree ring corresponds to one year. Tree: METHUSELAH WALK, PILO;
  Location: CALIFORNIA, GT BASIN B C PINE 2805M, 3726-11810; 
  Author: DONALD A. GRAYBILL, 1980.  
}
\source{
  Time Series Data Library:
  \url{http://www-personal.buseco.monash.edu.au/~hyndman/TSDL/}
}
\keyword{datasets}
