% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-package.R
\docType{package}
\name{ts-package}
\alias{ts}
\alias{ts-package}
\title{ts: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaborcsardi/ts}
  \item Report bugs at \url{https://github.com/gaborcsardi/ts/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
