% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-package.R
\docType{package}
\name{ts-package}
\alias{ts}
\alias{ts-package}
\title{ts: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
