% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-package.R
\docType{package}
\name{ts-package}
\alias{ts}
\alias{ts-package}
\title{ts: Tree-Sitter Parsing Tools}
\section{Installed ts parsers}{
\eval{ts:::format_rd_parser_list(ts:::ts_list_parsers())}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaborcsardi/ts}
  \item \url{https://gaborcsardi.github.io/ts/}
  \item Report bugs at \url{https://github.com/gaborcsardi/ts/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
