% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-list-parsers.R
\name{ts_list_parsers}
\alias{ts_list_parsers}
\title{List installed tree-sitter parsers}
\usage{
ts_list_parsers(lib_path = .libPaths())
}
\arguments{
\item{lib_path}{Library paths to search for installed packages.
Default is \code{\link[base:libPaths]{base::.libPaths()}}.}
}
\value{
A data frame with columns:
\itemize{
\item \code{package}: character, the name of the package.
\item \code{version}: character, the version of the package.
\item \code{title}: character, the title of the package.
\item \code{library}: character, the library path where the package is installed.
\item \code{loaded}: logical, whether the package is currently loaded.
}
}
\description{
List installed tree-sitter parsers
}
\examples{
ts_list_parsers()
}
