% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-list-parsers.R
\name{ts_list_parsers}
\alias{ts_list_parsers}
\title{List installed tree-sitter parsers}
\usage{
ts_list_parsers(lib_path = .libPaths())
}
\arguments{
\item{lib_path}{Library paths to search for installed packages.
Default is \code{\link[base:.libPaths]{base::.libPaths()}}.}
}
\value{
A data frame with columns:
\itemize{
\item \code{package}: character, the name of the package.
\item \code{version}: character, the version of the package.
\item \code{title}: character, the title of the package.
\item \code{library}: character, the library path where the package is installed.
\item \code{loaded}: logical, whether the package is currently loaded.
}
}
\description{
The ts package contains a common interface to several tree-sitter
parsers, implemented in other R packages. \code{ts_list_parsers()} lists
the available parsers installed in the system.
}
\details{
To see tree-sitter parser packages that are available on CRAN, but not
installed on your system, see the packages that depend on ts and have a
name with a 'ts' prefix.

Here is an example that includes all tree-sitter parsers at this time:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_list_parsers()
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #999999;"># A data frame: 2 × 5</span>                                                           
#>   package version    title           library                    loaded          
#> <span style="color: #c2c2c2;">*</span> <span style="font-style: italic;color: #999999;">&lt;chr&gt;</span>   <span style="font-style: italic;color: #999999;">&lt;chr&gt;</span>      <span style="font-style: italic;color: #999999;">&lt;chr&gt;</span>           <span style="font-style: italic;color: #999999;">&lt;chr&gt;</span>                      <span style="font-style: italic;color: #999999;">&lt;lgl&gt;</span>           
#> <span style="color: #c2c2c2;">1</span> tsjsonc 0.0.0.9000 Edit JSON Files /Users/gaborcsardi/Librar… FALSE           
#> <span style="color: #c2c2c2;">2</span> tstoml  0.0.0.9000 Edit TOML Files /Users/gaborcsardi/Librar… FALSE           
</pre></div>
}}
}
\examples{
ts_list_parsers()
}
