% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sitter.R
\name{ts_sexpr}
\alias{ts_sexpr}
\title{Show the syntax tree structure of a file or string}
\usage{
ts_sexpr(language, file = NULL, text = NULL, ranges = NULL)
}
\arguments{
\item{language}{Language of the file or string, a \code{ts_language} object.}

\item{file}{Path of a file. Use either \code{file} or \code{text}.}

\item{text}{String. Use either \code{file} or \code{text}, but not both.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}
}
\description{
Show the syntax tree structure of a file or string
}
\examples{
# TODO
}
