% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket-ts-tree.R
\name{ts_tree-brackets}
\alias{ts_tree-brackets}
\alias{[.ts_tree}
\title{Convert ts_tree object to a data frame}
\usage{
\method{[}{ts_tree}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{\eval{tsitter:::doc_insert("tsitter::ts_tree_brackets_param_x")}}

\item{i, j}{\eval{tsitter:::doc_insert("tsitter::ts_tree_brackets_param_ij")}}

\item{drop}{\eval{tsitter:::doc_insert("tsitter::ts_tree_brackets_param_drop")}}
}
\value{
\eval{tsitter:::doc_insert("tsitter::ts_tree_brackets_return")}
}
\description{
\eval{tsitter:::doc_insert("ts_tree_brackets_description")}
}
\details{
\eval{tsitter:::doc_insert("tsitter::ts_tree_brackets_details")}
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Create a parse tree with tsjsonc -------------------------------------
tree <- tsjsonc::ts_parse_jsonc('{"foo": 42, "bar": [1, 2, 3]}')

tree

tree[]
\dontshow{\}) # examplesIf}
}
\seealso{
\eval{tsitter:::doc_seealso("[")}

Other ts_tree exploration:
\code{\link[=ts_tree_ast]{ts_tree_ast()}},
\code{\link[=ts_tree_dom]{ts_tree_dom()}},
\code{\link[=ts_tree_query]{ts_tree_query()}},
\code{\link[=ts_tree_sexpr]{ts_tree_sexpr()}}
}
\concept{ts_tree exploration}
