% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-ast.R
\name{ts_tree_ast}
\alias{ts_tree_ast}
\title{Show the annotated syntax tree tree-sitter tree}
\usage{
ts_tree_ast(tree)
}
\arguments{
\item{tree}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_read]{ts_tree_read()}}.}
}
\description{
Show the annotated syntax tree tree-sitter tree
}
\examples{
# TODO
}
