% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-delete.R
\name{ts_tree_delete}
\alias{ts_tree_delete}
\title{TODO}
\usage{
ts_tree_delete(tree, ...)
}
\arguments{
\item{tree}{A \code{ts_tree} object.}

\item{...}{Extra arguments for methods.}
}
\description{
TODO
}
