% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-delete.R
\name{ts_tree_delete}
\alias{ts_tree_delete}
\title{Delete selected elements from a tree-sitter tree}
\usage{
ts_tree_delete(tree, ...)
}
\arguments{
\item{tree}{\eval{tsitter:::doc_insert("tsitter::ts_tree_delete_param_tree")}}

\item{...}{Extra arguments for methods.}
}
\value{
\eval{tsitter:::doc_insert("tsitter::ts_tree_delete_return")}
}
\description{
\eval{tsitter:::doc_insert("ts_tree_delete_description")}

\eval{tsitter:::format_rd_parser_list(tsitter:::ts_list_parsers(), "ts_tree_delete")}
}
\details{
\eval{tsitter:::doc_insert("tsitter::ts_tree_delete_details")}
\eval{tsitter:::doc_extra()}
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Create a parse tree with tsjsonc -------------------------------------
tree <- tsjsonc::ts_parse_jsonc(
  "{ \"a\": //comment\ntrue, \"b\": [1, 2, 3] }"
)

tree

tree \%>\% ts_tree_select("a")

tree \%>\% ts_tree_select("a") \%>\% ts_tree_delete()
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("tstoml", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Create a parse tree with tstoml --------------------------------------
tree <- tstoml::ts_parse_toml(r"(
  [servers]
  alpha = { ip = "127.0.0.1", dc = "eqdc10" }
  beta = { ip = "127.0.0.2", dc = "eqdc20" }
)")

tree

tree \%>\% ts_tree_select("servers", TRUE, "dc")
tree \%>\% ts_tree_select("servers", TRUE, "dc") \%>\% ts_tree_delete()
\dontshow{\}) # examplesIf}
}
\seealso{
\eval{tsitter:::doc_seealso("ts_tree_delete")}

Other ts_tree generics:
\code{\link[=[[.ts_tree]{[[.ts_tree()}},
\code{\link[=[[<-.ts_tree]{[[<-.ts_tree()}},
\code{\link[=format.ts_tree]{format.ts_tree()}},
\code{\link[=print.ts_tree]{print.ts_tree()}},
\code{\link{select-set}},
\code{\link[=ts_tree_ast]{ts_tree_ast()}},
\code{\link[=ts_tree_dom]{ts_tree_dom()}},
\code{\link[=ts_tree_format]{ts_tree_format()}},
\code{\link[=ts_tree_insert]{ts_tree_insert()}},
\code{\link[=ts_tree_new]{ts_tree_new()}},
\code{\link[=ts_tree_query]{ts_tree_query()}},
\code{\link[=ts_tree_select]{ts_tree_select()}},
\code{\link[=ts_tree_sexpr]{ts_tree_sexpr()}},
\code{\link[=ts_tree_unserialize]{ts_tree_unserialize()}},
\code{\link[=ts_tree_update]{ts_tree_update()}},
\code{\link[=ts_tree_write]{ts_tree_write()}}
}
\concept{ts_tree generics}
