% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-dom.R
\name{ts_tree_dom}
\alias{ts_tree_dom}
\title{Print the document object model (DOM) of a tree-sitter tree}
\usage{
ts_tree_dom(tree)
}
\arguments{
\item{tree}{A \code{ts_tree} object.}
}
\value{
Character vector, the formatted annotated syntax tree, line by
line. It has class \link[cli:tree]{cli_tree}, from the cli package. It
may contain ANSI escape sequences for coloring and hyperlinks.
}
\description{
\code{ts_tree_dom()} prints the document object model (DOM) tree of a ts_tree
object. This tree only includes semantic elements. E.g. for a JSON(C)
document it includes objects, arrays and various value types, but not
the syntax elements like brackets, commas or colons.
}
\details{
\subsection{The ts and ts* packages:}{

Language implementations may override the default \code{ts_tree_ast()} method,
to provide language-specific features. Make sure you read the correct
documentation for the language you are using.
}

\subsection{The syntax tree and the DOM tree}{

See \code{\link[=ts_tree_ast]{ts_tree_ast()}} for the complete tree-sitter syntax tree that
includes all nodes, including syntax elements like brackets and commas.
}
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{tree <- tsjsonc::ts_parse_jsonc('\{"foo": 42, "bar": [1, 2, 3]\}')
ts_tree_ast(tree)
#> document (1)                   1|
#> \\-object (2)                    |
#>   +-\{ (3)                       |\{
#>   +-pair (4)                    |
#>   | +-string (5)                |
#>   | | +-" (6)                   | "
#>   | | +-string_content (7)      |  foo
#>   | | \\-" (8)                   |     "
#>   | +-: (9)                     |      :
#>   | \\-number (10)               |        42
#>   +-, (11)                      |          ,
#>   +-pair (12)                   |
#>   | +-string (13)               |
#>   | | +-" (14)                  |            "
#>   | | +-string_content (15)     |             bar
#>   | | \\-" (16)                  |                "
#>   | +-: (17)                    |                 :
#>   | \\-array (18)                |
#>   |   +-[ (19)                  |                   [
#>   |   +-number (20)             |                    1
#>   |   +-, (21)                  |                     ,
#>   |   +-number (22)             |                       2
#>   |   +-, (23)                  |                        ,
#>   |   +-number (24)             |                          3
#>   |   \\-] (25)                  |                           ]
#>   \\-\} (26)                      |                            \}
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_dom(tree)
#> document (1)
#> \\-object (2)
#>   +-number (10) # foo
#>   \\-array (18) # bar
#>     +-number (20)
#>     +-number (22)
#>     \\-number (24)
}\if{html}{\out{</div>}}
}

\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# see the output above
tree <- tsjsonc::ts_parse_jsonc('{"foo": 42, "bar": [1, 2, 3]}')
tree
ts_tree_ast(tree)
ts_tree_dom(tree)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=ts_tree_ast]{ts_tree_ast()}} to show the annotated syntax tree of a
ts_tree object.

Other ts_tree functions: 
\code{\link{ts_tree_ast}()}
}
\concept{ts_tree functions}
