% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-dom.R
\name{ts_tree_dom}
\alias{ts_tree_dom}
\title{Print the document object model (DOM) of a tree-sitter tree}
\usage{
ts_tree_dom(tree)
}
\arguments{
\item{tree}{\eval{ts:::doc_insert("ts::ts_tree_dom_param_tree")}}
}
\value{
\eval{ts:::doc_insert("ts::ts_tree_dom_return")}
}
\description{
\eval{ts:::doc_insert("ts_tree_dom_description")}

\eval{ts:::format_rd_parser_list(ts:::ts_list_parsers(), "ts_tree_dom")}
}
\details{
\eval{ts:::doc_insert("ts::ts_tree_dom_details")}
\eval{ts:::doc_extra()}
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Create a parse tree with tsjsonc -------------------------------------
tree <- tsjsonc::ts_parse_jsonc('{"foo": 42, "bar": [1, 2, 3]}')

tree

ts_tree_ast(tree)

ts_tree_dom(tree)
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("tstoml", quietly = TRUE)) withAutoprint(\{ # examplesIf}

# Create a parse tree with tstoml --------------------------------------
tree <- tstoml::ts_parse_toml(r"(
  title = "TOML Example"
  [owner]
  name = "Tom Preston-Werner"
  dob = 1979-05-27T07:32:00-08:00
)")

tree

ts_tree_ast(tree)

ts_tree_dom(tree)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=ts_tree_ast]{ts_tree_ast()}} to show the annotated syntax tree of a
ts_tree object.

\eval{ts:::doc_seealso("ts_tree_dom")}

Other ts_tree exploration: 
\code{\link{ts_tree-brackets}},
\code{\link{ts_tree_ast}()},
\code{\link{ts_tree_query}()},
\code{\link{ts_tree_sexpr}()}

Other ts_tree generics: 
\code{\link{[[.ts_tree}()},
\code{\link{[[<-.ts_tree}()},
\code{\link{format.ts_tree}()},
\code{\link{print.ts_tree}()},
\code{\link{select-set}},
\code{\link{ts_tree_ast}()},
\code{\link{ts_tree_delete}()},
\code{\link{ts_tree_format}()},
\code{\link{ts_tree_insert}()},
\code{\link{ts_tree_new}()},
\code{\link{ts_tree_query}()},
\code{\link{ts_tree_select}()},
\code{\link{ts_tree_sexpr}()},
\code{\link{ts_tree_unserialize}()},
\code{\link{ts_tree_update}()},
\code{\link{ts_tree_write}()}
}
\concept{ts_tree exploration}
\concept{ts_tree generics}
