% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-dom.R
\name{ts_tree_dom}
\alias{ts_tree_dom}
\title{Print the document object model (DOM) of a tree-sitter tree}
\usage{
ts_tree_dom(tree)
}
\arguments{
\item{tree}{A \code{ts_tree} object.}
}
\value{
Character vector, the formatted annotated syntax tree, line by
line. It has class \link[cli:tree]{cli_tree}, from the cli package. It
may contain ANSI escape sequences for coloring and hyperlinks.
}
\description{
\code{ts_tree_dom()} prints the document object model (DOM) tree of a ts_tree
object. This tree only includes semantic elements. E.g. for a JSON(C)
document it includes objects, arrays and various value types, but not
the syntax elements like brackets, commas or colons.
}
\details{
\subsection{The syntax tree and the DOM tree}{

See \code{\link[tsitter:ts_tree_ast]{ts_tree_ast()}} for the complete
tree-sitter syntax tree that includes all nodes, including syntax
elements like brackets and commas.
}
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Create a parse tree with tsjsonc -------------------------------------
tree <- tsjsonc::ts_parse_jsonc('{"foo": 42, "bar": [1, 2, 3]}')

tree

ts_tree_ast(tree)

ts_tree_dom(tree)
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("tstoml", quietly = TRUE)) withAutoprint(\{ # examplesIf}

# Create a parse tree with tstoml --------------------------------------
tree <- tstoml::ts_parse_toml(r"(
  title = "TOML Example"
  [owner]
  name = "Tom Preston-Werner"
  dob = 1979-05-27T07:32:00-08:00
)")

tree

ts_tree_ast(tree)

ts_tree_dom(tree)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=ts_tree_ast]{ts_tree_ast()}} to show the annotated syntax tree of a
ts_tree object.

Other ts_tree exploration:
\code{\link[=[.ts_tree]{[.ts_tree()}},
\code{\link[=ts_tree_ast]{ts_tree_ast()}},
\code{\link[=ts_tree_query]{ts_tree_query()}},
\code{\link[=ts_tree_sexpr]{ts_tree_sexpr()}}

Other ts_tree generics:
\code{\link[=[[.ts_tree]{[[.ts_tree()}},
\code{\link[=[[<-.ts_tree]{[[<-.ts_tree()}},
\code{\link[=format.ts_tree]{format.ts_tree()}},
\code{\link[=print.ts_tree]{print.ts_tree()}},
\code{\link{select-set}},
\code{\link[=ts_tree_ast]{ts_tree_ast()}},
\code{\link[=ts_tree_delete]{ts_tree_delete()}},
\code{\link[=ts_tree_format]{ts_tree_format()}},
\code{\link[=ts_tree_insert]{ts_tree_insert()}},
\code{\link[=ts_tree_new]{ts_tree_new()}},
\code{\link[=ts_tree_query]{ts_tree_query()}},
\code{\link[=ts_tree_select]{ts_tree_select()}},
\code{\link[=ts_tree_sexpr]{ts_tree_sexpr()}},
\code{\link[=ts_tree_unserialize]{ts_tree_unserialize()}},
\code{\link[=ts_tree_update]{ts_tree_update()}},
\code{\link[=ts_tree_write]{ts_tree_write()}}
}
\concept{ts_tree exploration}
\concept{ts_tree generics}
