% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-format.R
\name{ts_tree_format}
\alias{ts_tree_format}
\title{Format a tree sitter tree for printing}
\usage{
ts_tree_format(tree, options, ...)
}
\arguments{
\item{tree}{A \code{ts_tree} object.}

\item{options}{A list of options for the formatting, see methods.}

\item{...}{Extra arguments for methods.}
}
\description{
Format a tree sitter tree for printing
}
