% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-insert.R
\name{ts_tree_insert}
\alias{ts_tree_insert}
\title{Insert a new element into a tree sitter tree}
\usage{
ts_tree_insert(tree, new, key, at, options, ...)
}
\arguments{
\item{tree}{A \code{ts_tree} object.}

\item{new}{The new element to insert.}

\item{key}{The key of the new element, if inserting into a keyed element.}

\item{at}{The position to insert the new element at. The interpretation of
this argument depends on the method that implements the insertion.}

\item{options}{A list of options for the insertion, see methods.}

\item{...}{Extra arguments for methods.}
}
\description{
TODO
}
