% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ts_tree_mark_selection1}
\alias{ts_tree_mark_selection1}
\title{TODO}
\usage{
ts_tree_mark_selection1(tree, node)
}
\arguments{
\item{tree}{Tree-sitter tree.}

\item{node}{Node id.}
}
\value{
Integer vector of node ids to highlight.
}
\description{
TODO
}
