% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-new.R
\name{ts_tree_new}
\alias{ts_tree_new}
\title{Create tree-sitter tree from file or string}
\usage{
ts_tree_new(
  language,
  file = NULL,
  text = NULL,
  ranges = NULL,
  fail_on_parse_error = TRUE,
  ...
)
}
\arguments{
\item{language}{Language of the file or string, a \code{ts_language} object,
e.g. \code{\link[tsjsonc:ts_language_jsonc]{tsjsonc::ts_language_jsonc()}}.}

\item{file}{Path of a file. Use either \code{file} or \code{text}, but not both.}

\item{text}{String. Use either \code{file} or \code{text}, but not both.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}

\item{fail_on_parse_error}{Logical, whether to error if there are
parse errors in the document. Default is \code{TRUE}.}

\item{...}{Additional arguments for methods.}
}
\value{
A data frame with one row per token, and columns:
\itemize{
\item \code{id}: integer, the id of the token. The (root) document node has id 1.
\item \code{parent}: integer, the id of the parent token. The root token has
parent \code{NA}
\item \code{field_name}: character, the field name of the token in its parent.
\item \code{type}: character, the type of the token.
\item \code{code}: character, the actual code of the token.
\item \code{start_byte}, \code{end_byte}: integer, the byte positions of the token
in the input.
\item \code{start_row}, \code{start_column}, \code{end_row}, \code{end_column}: integer, the
position of the token in the input.
\item \code{is_missing}: logical, whether the token is a missing token added by
the parser to recover from errors.
\item \code{has_error}: logical, whether the token has a parse error.
\item \code{children}: list of integer vectors, the ids of the children tokens.
\item \code{dom_type}: character, the type of the node in the DOM tree. See
\code{\link[=ts_tree_dom]{ts_tree_dom()}}. Nodes that are not part of the DOM tree have
\code{NA_character_} here.
\item \code{dom_children}: list of integer vectors, the ids of the children in the
DOM tree. See \code{\link[=ts_tree_dom]{ts_tree_dom()}}.
\item \code{dom_parent}: integer, the parent of the node in the DOM tree. See
\code{\link[=ts_tree_dom]{ts_tree_dom()}}. Nodes that are not part of the DOM tree and the
document node have have \code{NA_integer_} here.
}
}
\description{
This is the main function to create a tree-sitter parse tree, using a
parser from another package. Then the parse tree may be queried, edited,
formatted, written to file, etc. using ts_tree methods.
\subsection{Installed ts parsers}{

\eval{ts:::format_rd_parser_list(ts:::ts_list_parsers())}
}
}
\details{
A package that implements a tree-sitter parser provides a function that
creates a \code{ts_language} object for that parser. E.g.
\link[tsjsonc:tsjsonc-package]{tsjsonc} has \code{\link[tsjsonc:ts_language_jsonc]{tsjsonc::ts_language_jsonc()}}.
You need to use the returned \code{ts_language} object as the \code{language}
argument of this function.
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
json <- ts_tree_new(
  tsjsonc::ts_language_jsonc(),
  text = '{ "a": 1, "b": 2 }'
)
json
json |> ts_tree_format()
\dontshow{\}) # examplesIf}
}
\concept{ts_tree methods}
