% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-new.R
\name{ts_tree_new}
\alias{ts_tree_new}
\title{Create tree-sitter tree from file or string}
\usage{
ts_tree_new(
  language,
  file = NULL,
  text = NULL,
  ranges = NULL,
  fail_on_parse_error = TRUE,
  ...
)
}
\arguments{
\item{language}{\eval{ts:::doc_insert("ts::ts_tree_new_param_language", "ts")}}

\item{file}{\eval{ts:::doc_insert("ts::ts_tree_new_param_file", "ts")}}

\item{text}{\eval{ts:::doc_insert("ts::ts_tree_new_param_text", "ts")}}

\item{ranges}{\eval{ts:::doc_insert("ts::ts_tree_new_param_ranges", "ts")}}

\item{fail_on_parse_error}{\eval{ts:::doc_insert("ts::ts_tree_new_param_fail_on_parse_error", "ts")}}

\item{...}{Additional arguments for methods.}
}
\value{
\eval{ts:::doc_insert("ts::ts_tree_new_return", "ts")}
}
\description{
This is the main function to create a tree-sitter parse tree, using a
ts parser implemented in another package.
\eval{ts:::doc_insert("ts_tree_new_description", "ts")}

\eval{ts:::format_rd_parser_list(ts:::ts_list_parsers())}
}
\details{
A package that implements a tree-sitter parser provides a function that
creates a \code{ts_language} object for that parser. E.g.
\link[tsjsonc:tsjsonc-package]{tsjsonc} has \code{\link[tsjsonc:ts_language_jsonc]{tsjsonc::ts_language_jsonc()}}.
You need to use the returned \code{ts_language} object as the \code{language}
argument of \code{\link[ts:ts_tree_new]{ts_tree_new()}}.
\eval{ts:::doc_insert("ts_tree_new_details", "ts")}\if{html}{\out{
<p>
}}


\eval{ts:::doc_tabs("ts_tree_new_examples")}
\eval{ts:::doc_extra()}
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}

# JSONC example, needs the tsjsonc package -----------------------------
json <- ts_tree_new(
  tsjsonc::ts_language_jsonc(),
  text = '{ "a": 1, "b": 2 }'
)

json

json |> ts_tree_format()
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("tstoml", quietly = TRUE)) withAutoprint(\{ # examplesIf}

# TOML example, needs the tstoml package -------------------------------
toml <- ts_tree_new(
  tstoml::ts_language_toml(),
  text = '[section]\nkey = "value"\nnumber = 42\n'
)

toml

toml |> ts_tree_format()
\dontshow{\}) # examplesIf}
}
\seealso{
The tree-sitter parser packages typically include shortcuts to
create parse trees from strings and file, e.g.
\code{\link[tsjsonc:ts_parse_jsonc]{tsjsonc::ts_parse_jsonc()}} and \code{\link[tsjsonc:ts_parse_jsonc]{tsjsonc::ts_read_jsonc()}}.\if{html}{\out{
  <p>
}}


\eval{ts:::doc_seealso("ts_tree_new")}

Other \code{ts_tree} generics: 
\code{\link{[[.ts_tree}()},
\code{\link{[[<-.ts_tree}()},
\code{\link{format.ts_tree}()},
\code{\link{print.ts_tree}()},
\code{\link{select-set}},
\code{\link{ts_tree_ast}()},
\code{\link{ts_tree_delete}()},
\code{\link{ts_tree_dom}()},
\code{\link{ts_tree_format}()},
\code{\link{ts_tree_insert}()},
\code{\link{ts_tree_query}()},
\code{\link{ts_tree_select}()},
\code{\link{ts_tree_sexpr}()},
\code{\link{ts_tree_unserialize}()},
\code{\link{ts_tree_update}()},
\code{\link{ts_tree_write}()}
}
\concept{\code{ts_tree} generics}
