% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-new.R
\name{ts_tree_new}
\alias{ts_tree_new}
\title{Create tree-sitter tree from file or string}
\usage{
ts_tree_new(
  language,
  file = NULL,
  text = NULL,
  ranges = NULL,
  fail_on_parse_error = TRUE,
  ...
)
}
\arguments{
\item{language}{\eval{tsitter:::doc_insert("tsitter::ts_tree_new_param_language", "ts")}}

\item{file}{\eval{tsitter:::doc_insert("tsitter::ts_tree_new_param_file", "ts")}}

\item{text}{\eval{tsitter:::doc_insert("tsitter::ts_tree_new_param_text", "ts")}}

\item{ranges}{\eval{tsitter:::doc_insert("tsitter::ts_tree_new_param_ranges", "ts")}}

\item{fail_on_parse_error}{\eval{tsitter:::doc_insert("tsitter::ts_tree_new_param_fail_on_parse_error", "ts")}}

\item{...}{Additional arguments for methods.}
}
\value{
\eval{tsitter:::doc_insert("tsitter::ts_tree_new_return", "ts")}
}
\description{
This is the main function to create a tree-sitter parse tree, using a
ts parser implemented in another package.
\eval{tsitter:::doc_insert("ts_tree_new_description")}

\eval{tsitter:::format_rd_parser_list(tsitter:::ts_list_parsers())}
}
\details{
A package that implements a tree-sitter parser provides a function that
creates a \code{ts_language} object for that parser. E.g.
\link[tsjsonc:tsjsonc-package]{tsjsonc} has \code{\link[tsjsonc:ts_language_jsonc]{tsjsonc::ts_language_jsonc()}}.
You need to use the returned \code{ts_language} object as the \code{language}
argument of \code{\link[tsitter:ts_tree_new]{ts_tree_new()}}.

\eval{tsitter:::doc_insert("ts_tree_new_details")}
\eval{tsitter:::doc_extra()}
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}

# JSONC example, needs the tsjsonc package -----------------------------
json <- ts_tree_new(
  tsjsonc::ts_language_jsonc(),
  text = '{ "a": 1, "b": 2 }'
)

json

json \%>\% ts_tree_format()
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("tstoml", quietly = TRUE)) withAutoprint(\{ # examplesIf}

# TOML example, needs the tstoml package -------------------------------
toml <- ts_tree_new(
  tstoml::ts_language_toml(),
  text = '[section]\nkey = "value"\nnumber = 42\n'
)

toml

toml \%>\% ts_tree_format()
\dontshow{\}) # examplesIf}
}
\seealso{
The tree-sitter parser packages typically include shortcuts to
create parse trees from strings and file, e.g.
\code{\link[tsjsonc:ts_parse_jsonc]{tsjsonc::ts_parse_jsonc()}} and \code{\link[tsjsonc:ts_read_jsonc]{tsjsonc::ts_read_jsonc()}}.

\eval{tsitter:::doc_seealso("ts_tree_new")}

Other ts_tree generics:
\code{\link[=[[.ts_tree]{[[.ts_tree()}},
\code{\link[=[[<-.ts_tree]{[[<-.ts_tree()}},
\code{\link[=format.ts_tree]{format.ts_tree()}},
\code{\link[=print.ts_tree]{print.ts_tree()}},
\code{\link{select-set}},
\code{\link[=ts_tree_ast]{ts_tree_ast()}},
\code{\link[=ts_tree_delete]{ts_tree_delete()}},
\code{\link[=ts_tree_dom]{ts_tree_dom()}},
\code{\link[=ts_tree_format]{ts_tree_format()}},
\code{\link[=ts_tree_insert]{ts_tree_insert()}},
\code{\link[=ts_tree_query]{ts_tree_query()}},
\code{\link[=ts_tree_select]{ts_tree_select()}},
\code{\link[=ts_tree_sexpr]{ts_tree_sexpr()}},
\code{\link[=ts_tree_unserialize]{ts_tree_unserialize()}},
\code{\link[=ts_tree_update]{ts_tree_update()}},
\code{\link[=ts_tree_write]{ts_tree_write()}}
}
\concept{ts_tree generics}
