% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-query.R
\name{ts_tree_query}
\alias{ts_tree_query}
\title{Run tree-sitter queries on tree-sitter trees}
\usage{
ts_tree_query(tree, query)
}
\arguments{
\item{tree}{\eval{tsitter:::doc_insert("tsitter::ts_tree_query_param_tree")}}

\item{query}{\eval{tsitter:::doc_insert("tsitter::ts_tree_query_param_query")}}
}
\value{
\eval{tsitter:::doc_insert("tsitter::ts_tree_query_return")}
}
\description{
\eval{tsitter:::doc_insert("ts_tree_query_description")}
\eval{tsitter:::format_rd_parser_list(tsitter:::ts_list_parsers(), "ts_tree_query")}
}
\details{
\eval{tsitter:::doc_insert("tsitter::ts_tree_query_details")}
\eval{tsitter:::doc_tabs("ts_tree_query_details_examples")}
\eval{tsitter:::doc_extra()}
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Select all numbers in a JSONC document ------------------------------------
json <- tsjsonc::ts_parse_jsonc(
  '{ "a": 1, "b": [10, 20, 30], "c": { "c1": true, "c2": 100 } }'
)
json \%>\% ts_tree_query("(number) @number")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=ts_tree_select]{ts_tree_select()}} to select the nodes matching a query.

\eval{tsitter:::doc_seealso("ts_tree_query")}

Other ts_tree exploration:
\code{\link{ts_tree-brackets}},
\code{\link[=ts_tree_ast]{ts_tree_ast()}},
\code{\link[=ts_tree_dom]{ts_tree_dom()}},
\code{\link[=ts_tree_sexpr]{ts_tree_sexpr()}}

Other ts_tree generics:
\code{\link[=[[.ts_tree]{[[.ts_tree()}},
\code{\link[=[[<-.ts_tree]{[[<-.ts_tree()}},
\code{\link[=format.ts_tree]{format.ts_tree()}},
\code{\link[=print.ts_tree]{print.ts_tree()}},
\code{\link{select-set}},
\code{\link[=ts_tree_ast]{ts_tree_ast()}},
\code{\link[=ts_tree_delete]{ts_tree_delete()}},
\code{\link[=ts_tree_dom]{ts_tree_dom()}},
\code{\link[=ts_tree_format]{ts_tree_format()}},
\code{\link[=ts_tree_insert]{ts_tree_insert()}},
\code{\link[=ts_tree_new]{ts_tree_new()}},
\code{\link[=ts_tree_select]{ts_tree_select()}},
\code{\link[=ts_tree_sexpr]{ts_tree_sexpr()}},
\code{\link[=ts_tree_unserialize]{ts_tree_unserialize()}},
\code{\link[=ts_tree_update]{ts_tree_update()}},
\code{\link[=ts_tree_write]{ts_tree_write()}}
}
\concept{ts_tree exploration}
\concept{ts_tree generics}
