% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select}
\alias{ts_tree_select}
\title{Select parts of a tree-sitter tree}
\usage{
ts_tree_select(tree, ..., refine = FALSE)
}
\arguments{
\item{tree}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_new]{ts_tree_new()}}.}

\item{...}{Selection expressions, see details.}

\item{refine}{Logical, whether to refine the current selection or start
a new selection.}
}
\value{
A \code{ts_tree} object with the selected parts.
}
\description{
TODO: include links to methods here dynamically?
}
