% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select}
\alias{ts_tree_select}
\title{Select elements of a tree-sitter tree}
\usage{
ts_tree_select(tree, ..., refine = FALSE)
}
\arguments{
\item{tree}{\eval{ts:::doc_insert("ts_tree_select_param_tree")}}

\item{...}{\eval{ts:::doc_insert("ts_tree_select_param_dots")}}

\item{refine}{\eval{ts:::doc_insert("ts_tree_select_param_refine")}}
}
\value{
\eval{ts:::doc_insert("ts::ts_tree_select_return")}
}
\description{
\eval{ts:::doc_insert("ts::ts_tree_select_description")}

\eval{ts:::format_rd_parser_list(ts:::ts_list_parsers(), "ts_tree_select")}
}
\details{
\eval{ts:::doc_insert("ts::ts_tree_select_details")}
\eval{ts:::doc_extra()}
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# ----------------------------------------------------------------------
# Create a JSONC tree, needs the tsjsonc package
json <- ts_tree_new(
  tsjsonc::ts_language_jsonc(),
  text = '{ "a": 1, "b": 2, "c": { "d": 3, "e": 4 } }'
)

json |> ts_tree_select("c", "d")
\dontshow{\}) # examplesIf}
\dontshow{if (requireNamespace("tstoml", quietly = TRUE)) withAutoprint(\{ # examplesIf}

# ----------------------------------------------------------------------
# Create a TOML tree, needs the tstoml package
toml <- ts_tree_new(
  tstoml::ts_language_toml(),
  text = tstoml::toml_example_text()
)

toml |> ts_tree_select("servers", TRUE, "ip")
\dontshow{\}) # examplesIf}
}
\seealso{
\eval{ts:::doc_seealso("ts_tree_select")}

Other \code{ts_tree} generics: 
\code{\link{[[.ts_tree}()},
\code{\link{[[<-.ts_tree}()},
\code{\link{format.ts_tree}()},
\code{\link{print.ts_tree}()},
\code{\link{select-set}},
\code{\link{ts_tree_ast}()},
\code{\link{ts_tree_delete}()},
\code{\link{ts_tree_dom}()},
\code{\link{ts_tree_format}()},
\code{\link{ts_tree_insert}()},
\code{\link{ts_tree_new}()},
\code{\link{ts_tree_query}()},
\code{\link{ts_tree_sexpr}()},
\code{\link{ts_tree_unserialize}()},
\code{\link{ts_tree_update}()},
\code{\link{ts_tree_write}()}
}
\concept{\code{ts_tree} generics}
