% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select}
\alias{ts_tree_select}
\title{Select parts of a tree-sitter tree}
\usage{
ts_tree_select(tree, ...)
}
\arguments{
\item{tree}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_read]{ts_tree_read()}}.}

\item{...}{Selection expressions, see details.}
}
\description{
TODO: include links to methods here dynamically?
}
