% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select}
\alias{ts_tree_select}
\title{Select parts of a tree-sitter tree}
\usage{
ts_tree_select(tree, ..., refine = FALSE)
}
\arguments{
\item{tree}{\eval{ts:::doc_insert("ts_tree_select_param_tree")}}

\item{...}{\eval{ts:::doc_insert("ts_tree_select_param_dots")}}

\item{refine}{\eval{ts:::doc_insert("ts_tree_select_param_refine")}}
}
\value{
\eval{ts:::doc_insert("ts::ts_tree_select_return")}
}
\description{
\eval{ts:::doc_insert("ts::ts_tree_select_description")}
\subsection{Installed ts parsers}{

\eval{ts:::format_rd_parser_list(ts:::ts_list_parsers())}
}
}
\details{
\eval{ts:::doc_insert("ts::ts_tree_select_details")}
\eval{ts:::doc_extra()}
}
\examples{
\dontshow{if (requireNamespace("tsjsonc", quietly = TRUE)) withAutoprint(\{ # examplesIf}
json <- ts_tree_new(
  tsjsonc::ts_language_jsonc(),
  text = '{ "a": 1, "b": 2, "c": { "d": 3, "e": 4 } }'
)

json |> ts_tree_select("c", "d")
\dontshow{\}) # examplesIf}
}
