% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select1}
\alias{ts_tree_select1}
\title{TODO}
\usage{
ts_tree_select1(tree, node, slt)
}
\arguments{
\item{tree}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_new]{ts_tree_new()}}.}

\item{node}{Integer, the node id to select from.}

\item{slt}{A selector object, see details in \code{\link[=ts_tree_select]{ts_tree_select()}}.}
}
\description{
TODO
}
