% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-sexpr.R
\name{ts_tree_sexpr}
\alias{ts_tree_sexpr}
\title{Show the syntax tree structure of a file or string}
\usage{
ts_tree_sexpr(tree)
}
\arguments{
\item{tree}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_read]{ts_tree_read()}}.}
}
\description{
Show the syntax tree structure of a file or string
}
\examples{
# TODO
}
