% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-unserialize.R
\name{ts_tree_unserialize}
\alias{ts_tree_unserialize}
\title{TODO}
\usage{
ts_tree_unserialize(tree)
}
\arguments{
\item{tree}{A \code{ts_tree} object as returned by \code{\link[=ts_tree_read]{ts_tree_read()}}.}
}
\description{
TODO
}
