% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-update.R
\name{ts_tree_update}
\alias{ts_tree_update}
\title{TODO}
\usage{
ts_tree_update(tree, new, options, ...)
}
\arguments{
\item{tree}{A \code{ts_tree} object.}

\item{new}{R expression to serialize into a new element.}

\item{options}{A list of options for the update, see methods.}

\item{...}{Extra arguments for methods.}
}
\description{
TODO
}
