% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-write.R
\name{ts_tree_write}
\alias{ts_tree_write}
\title{Write a tree-sitter tree to a file}
\usage{
ts_tree_write(tree, file = NULL)
}
\arguments{
\item{tree}{\eval{ts:::doc_insert("ts::ts_tree_write_param_tree")}}

\item{file}{\eval{ts:::doc_insert("ts::ts_tree_write_param_file")}}
}
\value{
\eval{ts:::doc_insert("ts::ts_tree_write_return")}
}
\description{
\eval{ts:::doc_insert("ts_tree_write_description")}
}
\details{
\eval{ts:::doc_insert("ts::ts_tree_write_details")}
\eval{ts:::doc_extra()}
}
\examples{
\dontshow{if (Sys.getenv("IN_PKGDOWN") == "true") withAutoprint(\{ # examplesIf}
# Create a parse tree with tsjsonc -------------------------------------
tree <- tsjsonc::ts_parse_jsonc('{"foo": 42, "bar": [1, 2, 3]}')

# Format and write to file
tree |> ts_tree_format() |> ts_tree_write("example.json")
\dontshow{\}) # examplesIf}
}
\seealso{
\eval{ts:::doc_seealso("ts_tree_write")}

Other ts_tree generics: 
\code{\link{[[.ts_tree}()},
\code{\link{[[<-.ts_tree}()},
\code{\link{format.ts_tree}()},
\code{\link{print.ts_tree}()},
\code{\link{select-set}},
\code{\link{ts_tree_ast}()},
\code{\link{ts_tree_delete}()},
\code{\link{ts_tree_dom}()},
\code{\link{ts_tree_format}()},
\code{\link{ts_tree_insert}()},
\code{\link{ts_tree_new}()},
\code{\link{ts_tree_query}()},
\code{\link{ts_tree_select}()},
\code{\link{ts_tree_sexpr}()},
\code{\link{ts_tree_unserialize}()},
\code{\link{ts_tree_update}()}
}
\concept{ts_tree generics}
