% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-package.R
\docType{package}
\name{tsitter-package}
\alias{tsitter}
\alias{tsitter-package}
\title{tsitter: Tree-Sitter Parsing Tools}
\value{
Not applicable.
}
\description{
Common tree-sitter parsing tools for R. It is meant to be used by other packages that specialize in particular languages and file formats.
}
\details{
\eval{tsitter:::format_rd_parser_list(tsitter:::ts_list_parsers())}
}
\examples{
# See help("about", package = "tsitter")
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/tsitter}
  \item \url{https://r-lib.github.io/tsitter/}
  \item Report bugs at \url{https://github.com/r-lib/tsitter/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Gábor Csárdi \email{csardi.gabor@gmail.com}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
  \item Tree-sitter authors (Tree-sitter C library) [copyright holder]
}

}
\keyword{internal}
